Add some checks for uninitialized variables.

+++ SF/www/account/login.php
@@ -15,14 +15,14 @@ require_once('account.php');
 
 $Language->loadLanguageMsg('account/account');
 
-if (!session_issecure() && ($GLOBALS['sys_https_host'] != "")) {
+if (!session_issecure() && isset($GLOBALS['sys_https_host']) && ($GLOBALS['sys_https_host'] != "")) {
     //force use of SSL for login
     header('Location: https://'.$GLOBALS['sys_https_host'].'/account/login.php');
 }
 
 // ###### first check for valid login, if so, redirect
 
-if ($login) {
+if (isset($login) && $login) {
     list($success, $status) = session_login_valid($form_loginname,$form_pw);
     if ($success) {
 
@@ -75,14 +75,14 @@ if ($login) {
 	}
     }
 }
-if ($session_hash) {
+if (isset($session_hash) && $session_hash) {
 	//nuke their old session
 	session_cookie('session_hash','');
 	db_query("DELETE FROM session WHERE session_hash='$session_hash'");
 }
 $HTML->header(array('title'=>$Language->getText('account_login', 'title')));
 
-if ($login && !$success) {
+if (isset($login) && $login && !$success) {
 
     if ($status == 'P') {
 	echo "<P><B>".$Language->getText('account_login', 'pending_title')."</B>";
@@ -97,7 +97,7 @@ if ($login && !$success) {
     } //end else
 }
 
-if ($GLOBALS['sys_https_host']) {
+if (isset($GLOBALS['sys_https_host']) && $GLOBALS['sys_https_host']) {
     $form_url = "https://".$GLOBALS['sys_https_host'];
 } else {
     $form_url = "http://".$GLOBALS['sys_default_domain'];
@@ -105,17 +105,21 @@ if ($GLOBALS['sys_https_host']) {
 ?>
 
 <p>
-<h2><?php print $Language->getText('account_login', 'title', array($GLOBALS['sys_Name'])); ?>
-<?php print ($GLOBALS['sys_https_host'] != "" ? ' (Secure)':''); ?>
+<h2><?php 
+if(!isset($GLOBALS['sys_Name'])) {
+    $GLOBALS['sys_Name'] = "";
+ }
+print $Language->getText('account_login', 'title', array($GLOBALS['sys_Name'])); ?>
+<?php print ((isset($GLOBALS['sys_https_host']) && $GLOBALS['sys_https_host'] != "") ? ' (Secure)':''); ?>
 </h2>
 <p>
 <span class="highlight"><B><?php print $Language->getText('account_login', 'cookies'); ?>.</B></span>
 <P>
 <form action="<?php echo $form_url; ?>/account/login.php" method="post" name="form_login">
-<INPUT TYPE="HIDDEN" NAME="return_to" VALUE="<?php echo $return_to; ?>">
+<INPUT TYPE="HIDDEN" NAME="return_to" VALUE="<?php if (isset($return_to)) { echo $return_to; } ?>">
 <p>
 <?php print $Language->getText('account_login', 'name'); ?>:
-<br><input type="text" name="form_loginname" VALUE="<?php echo $form_loginname; ?>">
+<br><input type="text" name="form_loginname" VALUE="<?php if (isset($form_loginname)) { echo $form_loginname; } ?>">
 <p>
 <?php print $Language->getText('account_login', 'password'); ?>:
 <br><input type="password" name="form_pw">
@@ -124,7 +128,7 @@ if ($GLOBALS['sys_https_host']) {
 // Only show the stay in SSL mode if the server is SSL enabled
 // and it is not forced to operate in SSL mode
 // and the stay in SSL check box can be shown
-if ( $GLOBALS['sys_https_host'] != '' && $GLOBALS['sys_force_ssl'] == 0 &&
+if ( isset($GLOBALS['sys_https_host']) && $GLOBALS['sys_https_host'] != '' && $GLOBALS['sys_force_ssl'] == 0 &&
      $GLOBALS['sys_stay_in_ssl'] == 1 ) {
     echo '<INPUT TYPE="CHECKBOX" NAME="stay_in_ssl" VALUE="1" '.
     (((browser_is_ie() && browser_get_version() < '5.1') || !session_issecure()) ?'':'CHECKED').'>'.
